Fix conformability error when using randvars() option
When using option randvars(), the Cholesky matrix produced a
conformability error due to the fact that the part of the matrix
that comes from (potential) correlation between wages and
preferences was always filled even if wage prediction was turned
of (not to mention correlation between wages and preferences).
Now the submatrix containing potential correlation coefficients
is only filled when wage prediction is turned on.

Thanks to Aleksandra Anic for pointing this out.